flow-api/Jobs_GreenCoffee_TriggerUpdate/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2018.03.07=0
certifi=2018.4.16=py36_0
click=6.7=py36hec950be_0
flask=0.12.2=py36h5658096_0
intel-openmp=2018.0.0=8
itsdangerous=0.24=py36h49fbb8d_1
jinja2=2.10=py36hd36f9c5_0
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1=hb4e282d_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
markupsafe=1.0=py36h3a1e703_1
mkl=2018.0.2=1
mkl_fft=1.0.1=py36h917ab60_0
mkl_random=1.0.1=py36h78cc56f_0
ncurses=6.0=hd04f020_2
numpy=1.14.2=py36ha9ae307_1
openssl=1.0.2o=h26aff7b_0
pandas=0.22.0=py36h0a44026_0
pip=10.0.1=py36_0
python=3.6.5=hc167b69_1
python-dateutil=2.7.2=py36_0
pytz=2018.4=py36_0
readline=7.0=hc1231fa_4
setuptools=39.0.1=py36_0
six=1.11.0=py36h0e22d5e_1
sqlite=3.23.1=hf1716c9_0
tk=8.6.7=h35a86e2_3
werkzeug=0.14.1=py36_0
wheel=0.31.0=py36_0
xz=5.2.3=h727817e_4
zlib=1.2.11=hf3cbc9b_2
